+++ lib/Db/ShareMapper.php
@@ -25,10 +25,12 @@
 namespace OCA\Polls\Db;
 
 use OCA\Polls\Exceptions\ShareNotFoundException;
+use OCA\Polls\Migration\TableSchema;
 use OCA\Polls\Model\UserBase;
 use OCP\AppFramework\Db\DoesNotExistException;
 use OCP\AppFramework\Db\QBMapper;
 use OCP\DB\QueryBuilder\IQueryBuilder;
+use OCP\IConfig;
 use OCP\IDBConnection;
 
 /**
@@ -37,7 +39,10 @@ use OCP\IDBConnection;
 class ShareMapper extends QBMapper {
 	public const TABLE = Share::TABLE;
 
-	public function __construct(IDBConnection $db) {
+	public function __construct(
+		IDBConnection $db,
+		private IConfig $config,
+	) {
 		parent::__construct($db, Share::TABLE, Share::class);
 	}
 
@@ -47,16 +52,41 @@ class ShareMapper extends QBMapper {
 	 * @psalm-return array<array-key, Share>
 	 */
 	public function findByPoll(int $pollId): array {
+
 		$qb = $this->db->getQueryBuilder();
 
-		$qb->select('*')
-		   ->from($this->getTableName())
+		// get all column names from TableSchema
+		foreach (TableSchema::TABLES[Share::TABLE] as $column => $values) {
+			$selectColumns[] = 'p.' . $column;
+		}
+		// add vote counter
+		$selectColumns[] = $qb->func()->count('c1.id', 'voted');
+
+		// Build the following select (MySQL)
+		//
+		// SELECT p.*, COUNT(c1.user_id) as voted
+		// FROM oc_polls_share p
+		// LEFT JOIN oc_polls_votes c1
+		//   ON p.poll_id = c1.poll_id AND
+		// 	    p.user_id = c1.user_id
+		// GROUP BY p.poll_id, p.user_id
+		//
+		$qb->select($selectColumns)
+		   ->from($this->getTableName(), 'p')
 		   ->where(
-		   	$qb->expr()->eq('poll_id', $qb->createNamedParameter($pollId, IQueryBuilder::PARAM_INT))
-		   );
+		   	$qb->expr()->eq('p.poll_id', $qb->createNamedParameter($pollId, IQueryBuilder::PARAM_INT))
+		   )
+		   ->leftJoin('p', Vote::TABLE, 'c1',
+		   	$qb->expr()->andX(
+		   		$qb->expr()->eq('p.poll_id', 'c1.poll_id'),
+		   		$qb->expr()->eq('p.user_id', 'c1.user_id'),
+		   	)
+		   )
+			->groupBy('poll_id', 'user_id');
 
 		return $this->findEntities($qb);
 	}
+
 	/**
 	 * @throws \OCP\AppFramework\Db\DoesNotExistException if not found
 	 * @return Share[]
@@ -74,6 +104,18 @@ class ShareMapper extends QBMapper {
 
 		return $this->findEntities($qb);
 	}
+
+	public function countVotesByShare(Share $share): int {
+		$qb = $this->db->getQueryBuilder();
+
+		$qb->select('*')
+			->from($this->config->getSystemValue('dbtableprefix', 'oc_') . Vote::TABLE)
+			->where($qb->expr()->eq('poll_id', $qb->createNamedParameter($share->getPollId(), IQueryBuilder::PARAM_INT)))
+			->andWhere($qb->expr()->eq('user_id', $qb->createNamedParameter($share->getUserId(), IQueryBuilder::PARAM_STR)));
+
+		return count($this->findEntities($qb));
+	}
+
 	/**
 	 * @throws \OCP\AppFramework\Db\DoesNotExistException if not found
 	 * @return Share[]