+++ lib/Service/ShareService.php
@@ -32,6 +32,7 @@ use OCA\Polls\Event\ShareChangedRegistrationConstraintEvent;
 use OCA\Polls\Event\ShareCreateEvent;
 use OCA\Polls\Event\ShareDeletedEvent;
 use OCA\Polls\Event\ShareRegistrationEvent;
+use OCA\Polls\Event\ShareRevokedEvent;
 use OCA\Polls\Event\ShareTypeChangedEvent;
 use OCA\Polls\Exceptions\ForbiddenException;
 use OCA\Polls\Exceptions\InvalidShareTypeException;
@@ -321,6 +322,38 @@ class ShareService {
 		}
 
 		$this->eventDispatcher->dispatchTyped(new ShareDeletedEvent($share));
+
+		return $share->getToken();
+	}
+
+	/**
+	 * Revoke share
+	 */
+	public function revoke(Share $share = null, string $token = null): string {
+		if ($token) {
+			$share = $this->shareMapper->findByToken($token);
+		}
+		$this->acl->setPollId($share->getPollId(), Acl::PERMISSION_POLL_EDIT);
+
+		$share->setRevoked(time());
+		$this->shareMapper->update($share);
+		$this->eventDispatcher->dispatchTyped(new ShareRevokedEvent($share));
+
+		return $share->getToken();
+	}
+
+	/**
+	 * Re-revoke share
+	 */
+	public function reRevoke(Share $share = null, string $token = null): string {
+		if ($token) {
+			$share = $this->shareMapper->findByToken($token);
+		}
+		$this->acl->setPollId($share->getPollId(), Acl::PERMISSION_POLL_EDIT);
+
+		$share->setRevoked(0);
+		$this->shareMapper->update($share);
+		$this->eventDispatcher->dispatchTyped(new ShareCreateEvent($share));
 
 		return $share->getToken();
 	}