+++ src/js/components/SideBar/SideBarTabShare.vue
@@ -24,6 +24,7 @@
 	<div class="sidebar-share">
 		<SharesListUnsent class="shares unsent" />
 		<SharesList class="shares effective" />
+		<SharesListRevoked class="shares" />
 	</div>
 </template>
 
@@ -31,6 +32,7 @@
 import { mapState } from 'vuex'
 import SharesList from '../Shares/SharesList.vue'
 import SharesListUnsent from '../Shares/SharesListUnsent.vue'
+import SharesListRevoked from '../Shares/SharesListRevoked.vue'
 
 export default {
 	name: 'SideBarTabShare',
@@ -38,6 +40,7 @@ export default {
 	components: {
 		SharesList,
 		SharesListUnsent,
+		SharesListRevoked,
 	},
 
 	computed: {